reorganize project, modified Room db schema to add date field to Favorite entity
