+++ recipe-app/app/src/main/java/com/isaac/recipes/data/database/favorite/Favorite.kt
@@ -2,6 +2,7 @@ package com.isaac.recipes.data.database.favorite
 
 import androidx.room.Entity
 import androidx.room.PrimaryKey
+import java.util.*
 
 @Entity(tableName = "favorites_table")
 data class Favorite(
@@ -10,5 +11,6 @@ data class Favorite(
     val summary: String,
     val image: String,
     val ready_in_minutes: Int,
-    val servings: Int
+    val servings: Int,
+    val date_added: Date
 )
