+++ recipe-app/app/src/main/java/com/isaac/recipes/data/models/Recipe.kt
@@ -1,4 +1,4 @@
-package com.isaac.recipes.data
+package com.isaac.recipes.data.models
 
 import com.squareup.moshi.JsonClass
