+++ recipe-app/app/src/main/java/com/isaac/recipes/data/repos/FavoriteRepository.kt
@@ -1,4 +1,4 @@
-package com.isaac.recipes.data
+package com.isaac.recipes.data.repos
 
 import android.app.Application
 import androidx.lifecycle.LiveData
@@ -10,6 +10,7 @@ import com.isaac.recipes.data.database.relation.FavoriteWithDetails
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
+import java.util.*
 
 class FavoriteRepository(val app: Application) {
     private val db = AppDatabase.getDatabase(app)
@@ -26,7 +27,7 @@ class FavoriteRepository(val app: Application) {
     init {
         //dummy data in db
         CoroutineScope(Dispatchers.IO).launch {
-            favoriteDAO.insertFavorite(Favorite(1, "pasta", "my dinner", "pasta.png", 13, 1))
+            favoriteDAO.insertFavorite(Favorite(1, "pasta", "my dinner", "pasta.png", 13, 1, Date()))
             instructionDAO.insertInstruction(Instruction(recipe_id = 1, number = 1, step = "boil water"))
             instructionDAO.insertInstruction(Instruction(recipe_id = 1, number = 2, step = "salt water"))
             ingredientDAO.insertIngredient(Ingredient(recipe_id = 1, name = "pasta", amount = 1.0, unit = "pound"))
