+++ recipe-app/app/src/main/java/com/isaac/recipes/ui/search/SearchFragment.kt
@@ -1,20 +1,16 @@
 package com.isaac.recipes.ui.search
 
 import android.os.Bundle
-import android.util.Log
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import android.widget.Button
 import android.widget.EditText
 import androidx.fragment.app.Fragment
-import androidx.lifecycle.Observer
 import androidx.lifecycle.ViewModelProvider
 import androidx.navigation.NavController
 import androidx.navigation.Navigation
-import androidx.recyclerview.widget.RecyclerView
 import com.isaac.recipes.R
-import com.isaac.recipes.data.Recipe
 
 class SearchFragment : Fragment() {
